net/http: disable recently-introduced flaky test on Windows
Disable it until it's debugged so it doesn't hide other real
problems on Windows. The test was known to be unreliable
anyway (which is why it only needed 1 of 20 runs to pass), but
apparently it never passes on Windows. Figure out why later.

Update Issue 7634

LGTM=alex.brainman
R=adg, alex.brainman
CC=golang-codereviews
https://codereview.appspot.com/80110043